hours: Shorten STO maxAge to 1 hr
This brings things into parity with Carleton's hours cache expiry.

I can't think of a reason to have a longer time, to be honest. This
data is more ephemeral than we would like it to be, and since this
request is going out to a CDN, it doesn't really matter that we're
repeating it more often.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>