Change default server port to 80
This commit adjusts the HEALTHCHECK command and NODE_PORT default to
80 instead of 3001.

All of the reading I have done (fairly limited) seems to suggest that
it is a good idea to set the container up as if it were its own
dedicated server and not to do any monkey business.  Setting to a
non-standard HTTP server port (3001) is what I would consider monkey
business.

This will require updating the nginx configuration and deployment
configuration, and should be done carefully.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>