streams: Render date{To,From} in America/Chicago
This takes care of any translation necessary, and ensures that no
matter where the server is in the world, we will get consistent
results.

As long as moment-timezone knows our time zone, it also knows the
offset required to convert this to America/Chicago.  Thus, we interact
with the BMS API in times that are understood as being in CDT.  In
order to guarantee consistency, we generate moments and do our
math *first*, before converting to America/Chicago and then
formatting. This commit therefore resolves #89.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>