circle: Specify a version when setting up remote Docker engine
As suggested in [1], setup_remote_docker will default to the 17.09.0-ce
version of Docker.  Since we don't depend too strongly on the internals
of Docker being unchanged (and, more importantly, since [1] seems to
suggest the fix is to upgrade Docker) we can supply a version that
should build just fine.

If this builds greenly, I think this might be what's needed to get our
build working again.

[1]: https://support.circleci.com/hc/en-us/articles/360050934711-Docker-build-fails-with-EPERM-operation-not-permitted-copyfile-when-using-node-14-9-0-or-later-

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>