fix broken student jobs feed
The wp-json response now provides a much rougher response that requires
a lot of separate network requests that involve both json and html
parsing in order for us to deliver the full props for each job posting.

In short:
1. [json] wp-json paginated requests to get page urls (1-3 requests)
2. [html] detail page html requests, involves html parsing and returning
          json (easily 50+ requests)

We end up making multiple requests to the wp-json endpoint to build up
all of the top-level paginated urls, and then a request-per-post to get
the details.