server: Hook up sentry middleware
This is how we'll actually get insight into what's going on in the
request flow. I moved the two middleware functions into their own file.

- sentryRequestHandler: This middleware adds request data to the Sentry
  scope, so that we can see what's going on in the request.

- sentryTracingMiddleware: This middleware sets up a transaction for the
  request, so that we can see how long it took and what happened.

I also added a listener for the 'error' event on the app, so that we can
capture exceptions and send them to Sentry. This is a bit more involved
than the other two, because we have to add the request data to the
exception event ourselves.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>