nginx: Update configuration to match production
In preparation for a potential move to a new droplet, the nginx
configuration should probably match what it is in the real world.

This commit updates the nginx configuration to match what is as deployed
in the production droplet. Of note, we have both api.frogpond.tech and
frogpond.tech under the same SSL certificate now where previously (2018)
they were deployed under separate certificates.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>