sentry: Only log warn- and error-level messages to Sentry
The captureConsoleIntegration intercepts console messages and sends them
to Sentry. By default, it sends all messages. This commit changes the
set to just those levels that we probably care about and would have to
go digging for otherwise.

Of note, we do use Sentry.setupKoaErrorHandler() in server.ts, which
should negate the need for console.warn and console.error logs to be
sent anyways. But, this should drastically reduce the number of alerts
we get.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>